fix: separate reloadble functions into their own file
